feat: add postgres service - import pgx for running sql on testing database - setup script with testify.suite
