+++ test/config/TaskWithDefaultConnection.yaml
@@ -1,0 +1,17 @@
+---
+workers: 1
+logfile: TaskWithDefaultConnection.log
+
+connections:
+  - name: default
+    host: localhost
+    dbname: postgres
+    user: postgres
+    password: postgres
+    port: 5432
+
+tasks:
+  - id: 1
+    name: connect as dispatch by default
+    type: psql
+    command: \conninfo
