+++ test/expected/TaskWithDefaultConnection.log
@@ -3,5 +3,5 @@ Started at: 2000-01-01 00:00:00 +0000 UTC
 Ended at:   2000-01-01 00:00:00 +0000 UTC
 Error: 
 Output:
-hello world
+You are connected to database "postgres" as user "postgres" on host "localhost" (address "::1") at port "5432".
